backend/requirements-local.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - SQLite for local development
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
aiosqlite==0.19.0

# Authentication & Security  
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# File handling
aiofiles==23.2.0
python-magic==0.4.27

# Environment & Config
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2